Moving conventions to their own namespace
