+++ src/AutoMapper/Configuration/Conventions/DefaultName.cs
@@ -1,0 +1,6 @@
+namespace AutoMapper.Configuration.Conventions
+{
+    public class DefaultName : CaseSensitiveName
+    {
+    }
+}
