+++ src/AutoMapper/Configuration/Conventions/IChildMemberConfiguration.cs
@@ -1,0 +1,10 @@
+namespace AutoMapper.Configuration.Conventions
+{
+    using System;
+    using System.Collections.Generic;
+
+    public interface IChildMemberConfiguration
+    {
+        bool MapDestinationPropertyToSource(IProfileConfiguration options, TypeDetails sourceType, Type destType, string nameToSearch, LinkedList<IValueResolver> resolvers, IMemberConfiguration parent);
+    }
+}
