+++ src/AutoMapper/Configuration/Conventions/NameSplitMember.cs
@@ -1,0 +1,65 @@
+namespace AutoMapper.Configuration.Conventions
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
+    using System.Text.RegularExpressions;
+    using Internal;
+
+    public class NameSplitMember : IChildMemberConfiguration
+    {
+        public INamingConvention SourceMemberNamingConvention { get; set; }
+        public INamingConvention DestinationMemberNamingConvention { get; set; }
+
+        public NameSplitMember()
+        {
+            SourceMemberNamingConvention = new PascalCaseNamingConvention();
+            DestinationMemberNamingConvention = new PascalCaseNamingConvention();
+        }
+
+        public bool MapDestinationPropertyToSource(IProfileConfiguration options, TypeDetails sourceType, Type destType, string nameToSearch, LinkedList<IValueResolver> resolvers, IMemberConfiguration parent )
+        {
+            string[] matches = DestinationMemberNamingConvention.SplittingExpression
+                .Matches(nameToSearch)
+                .Cast<Match>()
+                .Select(m => SourceMemberNamingConvention.ReplaceValue(m))
+                .ToArray();
+            MemberInfo matchingMemberInfo = null;
+            for (int i = 1; i <= matches.Length; i++)
+            {
+                NameSnippet snippet = CreateNameSnippet(matches, i);
+
+                matchingMemberInfo = parent.NameMapper.GetMatchingMemberInfo(sourceType, destType, snippet.First);
+
+                if (matchingMemberInfo != null)
+                {
+                    resolvers.AddLast(matchingMemberInfo.ToMemberGetter());
+
+                    var foundMatch = parent.MapDestinationPropertyToSource(options, TypeMapFactory.GetTypeInfo(matchingMemberInfo.GetMemberType(), options), destType, snippet.Second, resolvers);
+
+                    if (!foundMatch)
+                        resolvers.RemoveLast();
+                    else
+                        break;
+                }
+            }
+            return matchingMemberInfo != null;
+        }
+        private NameSnippet CreateNameSnippet(IEnumerable<string> matches, int i)
+        {
+            var first = string.Join(SourceMemberNamingConvention.SeparatorCharacter, matches.Take(i).Select(s => SourceMemberNamingConvention.SplittingExpression.Replace(s, SourceMemberNamingConvention.ReplaceValue)).ToArray());
+            var second = string.Join(SourceMemberNamingConvention.SeparatorCharacter, matches.Skip(i).Select(s => SourceMemberNamingConvention.SplittingExpression.Replace(s, SourceMemberNamingConvention.ReplaceValue)).ToArray());
+            return new NameSnippet
+            {
+                First = first,
+                Second =second
+            };
+        }
+        private class NameSnippet
+        {
+            public string First { get; set; }
+            public string Second { get; set; }
+        }
+    }
+}
