+++ src/AutoMapper/Internal/ISourceToDestinationNameMapper.cs
@@ -3,6 +3,8 @@ using System.Reflection;
 
 namespace AutoMapper
 {
+    using Configuration.Conventions;
+
     public interface ISourceToDestinationNameMapper
     {
         MemberInfo GetMatchingMemberInfo(IGetTypeInfoMembers getTypeInfoMembers, TypeDetails typeInfo, Type destType, string nameToSearch);
