+++ src/AutoMapper/MapperConfiguration.cs
@@ -5,6 +5,7 @@ namespace AutoMapper
     using System.Collections.Concurrent;
     using System.Linq;
     using System.Reflection;
+    using Configuration.Conventions;
     using Internal;
     using Mappers;
     using QueryableExtensions;
