+++ src/UnitTests/MapToAttributeTest.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper.Mappers;
+using AutoMapper.Configuration.Conventions;
 using Should;
 using Xunit;
